Create the new `@SendablePropertyUnchecked` macro
