+++ Sources/SendablePropertyMacros/SendablePropertyMacro.swift
@@ -21,19 +21,20 @@ import SwiftSyntax
 import SwiftSyntaxBuilder
 import SwiftSyntaxMacros
 
-/// A macro that allows to make a property thread-safe keeping the `Sendable` conformance of the type.
+/// A macro that allows to make a property of a supported type thread-safe keeping the `Sendable` conformance of the type.
 public struct SendablePropertyMacro: PeerMacro {
-    private static let allowedTypes: Set<String> = ["Int", "UInt", "Int16", "UInt16", "Int32", "UInt32", "Int64", "UInt64", "Float", "Double", "Bool", "UnsafeRawPointer", "UnsafeMutableRawPointer", "OpaquePointer"]
-    
-    private static func peerPropertyName(for propertyName: String) -> String {
-        "_" + propertyName
-    }
-    
-    private static func checkPropertyType(for binding: PatternBindingSyntax, in declaration: some DeclSyntaxProtocol) throws {
-        guard let typeAnnotation = binding.typeAnnotation else {
-            return
-        }
-        guard let id = typeAnnotation.type.as(IdentifierTypeSyntax.self) else {
+    private static let allowedTypes: Set<String> = [
+        "Int", "UInt", "Int16", "UInt16", "Int32", "UInt32", "Int64", "UInt64", "Float", "Double", "Bool", "UnsafeRawPointer", "UnsafeMutableRawPointer", "UnsafePointer",
+        "UnsafeMutablePointer",
+    ]
+
+    private static func checkPropertyType(in declaration: some DeclSyntaxProtocol) throws {
+        guard let varDecl = declaration.as(VariableDeclSyntax.self),
+            let binding = varDecl.bindings.first,
+            let typeAnnotation = binding.typeAnnotation,
+            let id = typeAnnotation.type.as(IdentifierTypeSyntax.self)
+        else {
+            // Nothing to check.
             return
         }
 
@@ -46,6 +47,7 @@ public struct SendablePropertyMacro: PeerMacro {
         if typeName.contains("<") {
             typeName = String(typeName.prefix { $0 != "<" })
         }
+
         guard allowedTypes.contains(typeName) else {
             throw SendablePropertyError.notApplicableToType
         }
@@ -59,34 +61,8 @@ public struct SendablePropertyMacro: PeerMacro {
     public static func expansion(
         of node: SwiftSyntax.AttributeSyntax, providingPeersOf declaration: some SwiftSyntax.DeclSyntaxProtocol, in context: some SwiftSyntaxMacros.MacroExpansionContext
     ) throws -> [SwiftSyntax.DeclSyntax] {
-        guard let varDecl = declaration.as(VariableDeclSyntax.self),
-            let binding = varDecl.bindings.first,
-            let pattern = binding.pattern.as(IdentifierPatternSyntax.self)
-        else {
-            throw SendablePropertyError.onlyApplicableToVar
-        }
-
-        try checkPropertyType(for: binding, in: declaration)
-        
-        let propertyName = pattern.identifier.text
-        let hasInitializer = binding.initializer != nil
-        let initializerValue = binding.initializer?.value.description ?? "nil"
-
-        var genericTypeAnnotation = ""
-        if let typeAnnotation = binding.typeAnnotation {
-            let typeName = typeAnnotation.type.description.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
-            genericTypeAnnotation = "<\(typeName)\(hasInitializer ? "" : "?")>"
-        }
-
-        let accessLevel = varDecl.modifiers.first(where: { ["open", "public", "internal", "fileprivate", "private"].contains($0.name.text) })?.name.text ?? "internal"
-
-        // Create a peer property
-        let peerPropertyName = self.peerPropertyName(for: propertyName)
-        let peerProperty: DeclSyntax =
-            """
-            \(raw: accessLevel) let \(raw: peerPropertyName) = Synchronized\(raw: genericTypeAnnotation)(\(raw: initializerValue))
-            """
-        return [peerProperty]
+        try checkPropertyType(in: declaration)
+        return try SendablePropertyMacroUnchecked.expansion(of: node, providingPeersOf: declaration, in: context)
     }
 }
 
@@ -99,34 +75,7 @@ extension SendablePropertyMacro: AccessorMacro {
     public static func expansion(
         of node: SwiftSyntax.AttributeSyntax, providingAccessorsOf declaration: some SwiftSyntax.DeclSyntaxProtocol, in context: some SwiftSyntaxMacros.MacroExpansionContext
     ) throws -> [SwiftSyntax.AccessorDeclSyntax] {
-        guard let varDecl = declaration.as(VariableDeclSyntax.self),
-            let binding = varDecl.bindings.first,
-            let pattern = binding.pattern.as(IdentifierPatternSyntax.self)
-        else {
-            throw SendablePropertyError.onlyApplicableToVar
-        }
-
-        try checkPropertyType(for: binding, in: declaration)
-
-        let propertyName = pattern.identifier.text
-        let hasInitializer = binding.initializer != nil
-
-        // Replace the property with an accessor
-        let peerPropertyName = Self.peerPropertyName(for: propertyName)
-
-        let accessorGetter: AccessorDeclSyntax =
-            """
-            get {
-                \(raw: peerPropertyName).withLock { $0\(raw: hasInitializer ? "" : "!") }
-            }
-            """
-        let accessorSetter: AccessorDeclSyntax =
-            """
-            set {
-                \(raw: peerPropertyName).withLock { $0 = newValue }
-            }
-            """
-
-        return [accessorGetter, accessorSetter]
+        try checkPropertyType(in: declaration)
+        return try SendablePropertyMacroUnchecked.expansion(of: node, providingAccessorsOf: declaration, in: context)
     }
 }
